Separate uwsgi plugins build and storage for later recall
The changes to the installer.sh script and the addition of the uwsgi.yml file
in the docker-compose directory are intended to separate the build of the psgi
and http uwsgi plugins and store them in a location that can be accessed later
when running uwsgi. Previously, the plugins were being built and stored in the
/tmp directory, which would not be persisted after the container was stopped.
Storing the plugins in the /usr/local/bin directory allows them to be accessed
later when needed. The change to the web.conf file is intended to specify the
directory containing the plugins so that uwsgi can find them when it is run.