Improve logging and monitoring with updated logrotate configuration
As part of our ongoing efforts to improve the logging and monitoring of our
system, we decided to update the logrotate configuration for the
/var/log/uwsgi/app.log file.

Previously, we were only keeping one rotated log file, which meant that we were
only able to retain a limited amount of historical log data. To improve our
ability to track and troubleshoot issues, we decided to increase the number of
rotated log files to 7. This will allow us to retain a week's worth of log
data, giving us more visibility into the system's behavior over time.

In addition to increasing the number of rotated log files, we also decided to
change the compression utility used by logrotate. Previously, we were using
gzip, which is a widely supported compression format, but it is not always the
most efficient when it comes to compressing log files. After evaluating
different options, we decided to switch to tar and bzip2, using the highest
compression level (-9). This will allow us to achieve better compression
ratios, resulting in smaller compressed log files that take up less storage
space.

Finally, we also added the copytruncate directive to the logrotate
configuration. This tells logrotate to copy the log file and truncate the
original file to zero bytes, rather than physically moving or deleting the
file. This ensures that the log file continues to be written to, even after it
has been rotated, and it helps to prevent any disruptions to the logging
process.

Overall, these changes will help us to improve the efficiency and reliability
of our logging and monitoring system, and they will give us a better
understanding of the system's behavior over time.