Update logrotate config file and ownership
This commit updates the logrotate config file to compress logs using tar with
the -czf option instead of bzip2 with the -cjf9 option. This change was made
because using bzip2 on production resulted in errors.

In addition, this commit changes the ownership of the config file to root using
chown. This is necessary to ensure that the log rotation process has the
necessary permissions to access and modify the config file.